+++ 2Domain/QRCanvasViewer.cpp
@@ -7,18 +7,18 @@
 using namespace std;
 
 void QRCanvasViewer::validateD() {
-    dx = max(0., dx);
-    dy = max(0., dy);
+    dx = max(0.f, dx);
+    dy = max(0.f, dy);
     dx = min(dx, src->width() * (1 - 1/k));
     dy = min(dy, src->height() * (1 - 1/k));
 }
 void QRCanvasViewer::validateK() {
-    k = max(1., k);
+    k = max(1.f, k);
 }
 
-void QRCanvasViewer::zoom(Point center, double k_new) {
-    double dx_new = (center.x - src->width()/2) / k / zoomSensivity;
-    double dy_new = (center.y - src->height()/2) / k / zoomSensivity;
+void QRCanvasViewer::zoom(Point center, float k_new) {
+    float dx_new = (center.x - src->width()/2) / k / zoomSensivity;
+    float dy_new = (center.y - src->height()/2) / k / zoomSensivity;
     k *= k_new;
     validateK();
 
@@ -27,7 +27,7 @@ void QRCanvasViewer::zoom(Point center, double k_new) {
     validateD();
 }
 
-void QRCanvasViewer::move(double _dx, double _dy) {
+void QRCanvasViewer::move(float _dx, float _dy) {
     dx += _dx / moveSensivity;
     dy += _dy / moveSensivity;
     validateD();
