+++ 2Domain/QRCanvasViewer.h
@@ -11,17 +11,17 @@
 #include "../globalDefines.h"
 #include "2DStructures.h"
 
-const double zoomRatio = 1.1;
-const double zoomSensivity = 4;
-const double moveSensivity = 32;
+const float zoomRatio = 1.1;
+const float zoomSensivity = 4;
+const float moveSensivity = 32;
 
 class QRCanvasViewer {
 public:
     QRCanvasViewer(sptr<QImage> img = nullptr): src(img), k(1), dx(0), dy(0) {}
     void drawOn(sptr<QImage> img);
 
-    void zoom(Point, double);
-    void move(double, double);
+    void zoom(Point, float);
+    void move(float, float);
 
     void reset();
 private:
@@ -29,7 +29,7 @@ private:
     void validateK();
 
     sptr<QImage> src;
-    double dx, dy, k;
+    float dx, dy, k;
 };
 
 #endif //BIG3DFLUFFY_QRCANVASVIEWER_H
