+++ 3Domain/math/Functions.h
@@ -8,12 +8,12 @@
 #include "QRConstants.h"
 
 // todo wtf - not linking when impl is in cpp
-inline char sign(double x) {
+inline char sign(float x) {
     return x > QREPS? 1 : (x < -QREPS? -1 : 0);
 }
 
 // todo not ints
-inline int QRound(double x) {
+inline int QRound(float x) {
     int y = x;
     return x >= y + 0.5? y+1 : y;
 }
