+++ 3Domain/math/Transformer3D.cpp
@@ -6,7 +6,7 @@
 
 using namespace std;
 
-MoveTransformer3DCreator::MoveTransformer3DCreator(double dx, double dy, double dz) {
+MoveTransformer3DCreator::MoveTransformer3DCreator(float dx, float dy, float dz) {
     matrix = makeID();
     matrix[0][3] = dx, matrix[1][3] = dy, matrix[2][3] = dz;
 }
@@ -14,7 +14,7 @@ uptr<QRTransformer3D> MoveTransformer3DCreator::create() {
     return uptr<QRTransformer3D>(new Transformer3D(matrix));
 }
 
-ScaleTransformer3DCreator::ScaleTransformer3DCreator(double kx, double ky, double kz) {
+ScaleTransformer3DCreator::ScaleTransformer3DCreator(float kx, float ky, float kz) {
     matrix = makeID();
     matrix[0][0] = kx, matrix[1][1] = ky, matrix[2][2] = kz;
 }
@@ -23,7 +23,7 @@ uptr<QRTransformer3D> ScaleTransformer3DCreator::create() {
 }
 
 
-RotateTransformer3DCreator::RotateTransformer3DCreator(double dx, double dy, double dz) {
+RotateTransformer3DCreator::RotateTransformer3DCreator(float dx, float dy, float dz) {
     matrix = makeID();
     if (fabs(dx) > QREPS)
         matrix *= createOneRotateMatrix(dx, ox);
@@ -33,7 +33,7 @@ RotateTransformer3DCreator::RotateTransformer3DCreator(double dx, double dy, dou
         matrix *= createOneRotateMatrix(dz, oz);
 }
 
-Matrix3D RotateTransformer3DCreator::createOneRotateMatrix(double rad, axis ax) {
+Matrix3D RotateTransformer3DCreator::createOneRotateMatrix(float rad, axis ax) {
     Matrix3D matr = makeID();
     if (ax == ox) {
         matr[1][1] = cos(rad);
@@ -64,7 +64,7 @@ ProjectionTransformer3DCreator::ProjectionTransformer3DCreator(const Vector3D &v
     if (abs(v[1]) > QREPS) matrix[3][1] = -1/v[1] * PROJECTIVE_COEF;
     if (abs(v[2]) > QREPS) matrix[3][2] = -1/v[2] * PROJECTIVE_COEF;
 }
-ProjectionTransformer3DCreator::ProjectionTransformer3DCreator(double x, double y, double z)
+ProjectionTransformer3DCreator::ProjectionTransformer3DCreator(float x, float y, float z)
 :ProjectionTransformer3DCreator(Vector3D(x,y,z)) {}
