+++ 3Domain/math/Vector3D.cpp
@@ -11,23 +11,23 @@ Vector3D::Vector3D(double x, double y, double z) {
 Vector3D::Vector3D(double x, double y, double z, double p) {
     arr[0] = x, arr[1] = y, arr[2] = z, arr[3] = p;
 }
-Vector3D::Vector3D(const double _arr[4]) {
+Vector3D::Vector3D(const float _arr[4]) {
     arr[0] = _arr[0], arr[1] = _arr[1];
     arr[2] = _arr[2], arr[3] = _arr[3];
 }
 
-Vector3D::Vector3D(const std::initializer_list<double> &lst) {
+Vector3D::Vector3D(const std::initializer_list<float> &lst) {
     char i = 0;
     for (auto x: lst) {
         arr[i++] = x;
     }
 }
 
-double& Vector3D::operator[](int ind) {
+float& Vector3D::operator[](int ind) {
     return arr[ind];
 }
 
-const double& Vector3D::operator[](int ind) const {
+const float& Vector3D::operator[](int ind) const {
     return arr[ind];
 }
 
@@ -42,7 +42,7 @@ Vector3D& Vector3D::operator -=(const Vector3D &v) {
     return *this;
 }
 
-Vector3D& Vector3D::operator /=(double x) {
+Vector3D& Vector3D::operator /=(float x) {
     arr[0] /= x, arr[1] /= x;
     arr[2] /= x, arr[3] /= x;
     return *this;
@@ -71,19 +71,19 @@ Vector3D operator *(const Vector3D &a0, const Vector3D &b0) {
                      0});
 }
 
-Vector3D operator *(const Vector3D &a, double x) {
+Vector3D operator *(const Vector3D &a, float x) {
     return Vector3D({a.arr[0]*x, a.arr[1]*x, a.arr[2]*x, a.arr[3]*x});
 }
-Vector3D operator *(double x, const Vector3D &a) {
+Vector3D operator *(float x, const Vector3D &a) {
     return Vector3D({a.arr[0]*x, a.arr[1]*x, a.arr[2]*x, a.arr[3]*x});
 }
 
-Vector3D operator /(const Vector3D &a, double x) {
+Vector3D operator /(const Vector3D &a, float x) {
     return Vector3D({a.arr[0]/x, a.arr[1]/x, a.arr[2]/x, a.arr[3]/x});
 }
 
 Vector3D norm(const Vector3D &a) {
-    double x = a[3];
+    float x = a[3];
     if (fabs(x) < QREPS) return a;
     return a / x;
 }
@@ -96,23 +96,23 @@ void Vector3D::normSelf() {
 }
 
 Vector3D lenNorm(const Vector3D &a) {
-    double x = sqrt(a[0]*a[0]+a[1]*a[1]+a[2]*a[2] + a[3]*a[3]);
+    float x = sqrt(a[0]*a[0]+a[1]*a[1]+a[2]*a[2] + a[3]*a[3]);
     if (fabs(x) < QREPS) return a;
     return a / x;
 }
 void Vector3D::lenNormSelf() {
-    double x = sqrt(arr[0]*arr[0]+arr[1]*arr[1]+arr[2]*arr[2] + arr[3]*arr[3]);
+    float x = sqrt(arr[0]*arr[0]+arr[1]*arr[1]+arr[2]*arr[2] + arr[3]*arr[3]);
     arr[0] /= x;
     arr[1] /= x;
     arr[2] /= x;
     arr[3] /= x;
 }
 
-double vectorLen(const Vector3D &v) {
+float vectorLen(const Vector3D &v) {
     return sqrt(v[0]*v[0] + v[1]*v[1] + v[2]*v[2] + v[3]*v[3]);
 }
 
-double scalar(const Vector3D &a, const Vector3D &b) {
+float scalar(const Vector3D &a, const Vector3D &b) {
     return a[0]*b[0] + a[1]*b[1] + a[2]*b[2] + a[3]*b[3];
 }
 
@@ -121,7 +121,7 @@ bool isRightRotate(const Vector3D &a, const Vector3D &b, const Vector3D &c) {
     return sign(v1[0] * v2[1] - v1[1] * v2[0]) < 0;
 }
 
-double cos(const Vector3D &a, const Vector3D &b) {
+float cos(const Vector3D &a, const Vector3D &b) {
     return scalar(a,b) / vectorLen(a) / vectorLen(b);
 }
