+++ 3Domain/math/Vector3D.h
@@ -25,19 +25,19 @@ public:
     Vector3D();
     Vector3D(double, double, double);
     Vector3D(double, double, double, double);
-    explicit Vector3D(const double [4]);
-    Vector3D(const std::initializer_list<double> &);
+    explicit Vector3D(const float [4]);
+    Vector3D(const std::initializer_list<float> &);
 
     Vector3D& operator +=(const Vector3D&);
     Vector3D& operator -=(const Vector3D&);
-    Vector3D& operator /=(double);
-    inline double& operator[](int);
-    const double& operator[](int) const;
+    Vector3D& operator /=(float);
+    inline float& operator[](int);
+    const float& operator[](int) const;
 
     void normSelf();
     void lenNormSelf();
 
-    double arr[4] = {0,0,0,0};
+    float arr[4] = {0.f,0.f,0.f,0.f};
 };
 
 const Vector3D XVector(1,0,0,0);
@@ -47,18 +47,18 @@ const Vector3D ZVector(0,0,1,0);
 
 Vector3D norm(const Vector3D&); // manage 4th axis value, if 0 - does nothing
 Vector3D lenNorm(const Vector3D&);
-double vectorLen(const Vector3D&);
+float vectorLen(const Vector3D&);
 bool operator ==(const Vector3D&, const Vector3D&);
 bool operator !=(const Vector3D&, const Vector3D&);
 Vector3D operator +(const Vector3D&, const Vector3D&);
 Vector3D operator -(const Vector3D&, const Vector3D&);
 Vector3D operator *(const Vector3D&, const Vector3D&);  // 3dim-vector product, 4th is 0
-Vector3D operator *(const Vector3D&, double);
-Vector3D operator *(double, const Vector3D&);
-Vector3D operator /(const Vector3D&, double);
+Vector3D operator *(const Vector3D&, float);
+Vector3D operator *(float, const Vector3D&);
+Vector3D operator /(const Vector3D&, float);
 
-double scalar(const Vector3D &a, const Vector3D &b);
-double cos(const Vector3D &a, const Vector3D &b);
+float scalar(const Vector3D &a, const Vector3D &b);
+float cos(const Vector3D &a, const Vector3D &b);
 
 // for 2dim-vectors. others ignored
 bool isRightRotate(const Vector3D &a, const Vector3D &b, const Vector3D &c);
