+++ 3Domain/objects/components/impl/Point3D.cpp
@@ -6,4 +6,4 @@
 
 Point3D::Point3D(const Vector3D &_vec, QRPointStyle s): QRPoint3D(s), vec(_vec) {vec[3] = 0;}
 
-Point3D::Point3D(const double x, const double y, const double z, QRPointStyle s): QRPoint3D(s), vec(x, y, z, 0) {}
+Point3D::Point3D(const float x, const float y, const float z, QRPointStyle s): QRPoint3D(s), vec(x, y, z, 0) {}
