+++ 3Domain/objects/components/impl/Point3D.h
@@ -19,7 +19,7 @@ class Point3D: public QRPoint3D {
 public:
     Point3D() {}
     Point3D(const Vector3D &vec, QRPointStyle s = QRPointStyle());
-    Point3D(const double x, const double y, const double z, QRPointStyle s = QRPointStyle());
+    Point3D(const float x, const float y, const float z, QRPointStyle s = QRPointStyle());
     virtual sptr<QRObject3D> copy() {return sptr<QRPoint3D>(new Point3D(vec, getStyle()));}
 
     virtual const Vector3D getVector() const {return vec;}
