+++ 3Domain/objects/components/legacy/SimpleCamera3D.cpp
@@ -5,7 +5,7 @@
 #include "SimpleCamera3D.h"
 
 // todo get point already
-SimpleCamera3D::SimpleCamera3D(double w, double h, sptr<ProjectionImp>pr, const Vector3D &_origin,
+SimpleCamera3D::SimpleCamera3D(float w, float h, sptr<ProjectionImp>pr, const Vector3D &_origin,
                                const Vector3D &viewUp, const Vector3D &_bind)
         : QRCamera3D(w, h, pr), origin(new SimplePoint3D(_origin, _bind)), viewUpVector(viewUp) {
 }
