+++ globalDefines.h
@@ -11,8 +11,8 @@
 #define sptr std::shared_ptr
 #define uptr std::unique_ptr
 
-const double SCENE_DIST = 10;
-const double PROJECTIVE_COEF = 1/100;
+const float SCENE_DIST = 10;
+const float PROJECTIVE_COEF = 1/100;
 
 
 #endif //BIG3DFLUFFY_GLOBALDEFINES_H
